Bug fix: ReadGroupPerDuplexMolecularId b-depth was incorrect. (#30)
No current tool relies upon these values, so changes in performance is
expected.

The counts of the raw reads on the B-strand was incorrectly set as the
counts of raw reads on the A-strand.  Also, the tags used to store
these values have been updated to match fgbio's consensus tags.